Moved doLogin method to login plugin for reusability
